Separate macros into a separate file
