+++ sandbox/testinputs/groundtruth.cpp
@@ -1,17 +1,8 @@
 #include "groundtruth.h"
+
 #include "functions.hpp"
+#include "macros.h"
 
 #include <vector>
-
-#define GT_DEFINE(funcName)                                         \
-  float gt_##funcName(const std::vector<float> &in) {               \
-    return funcName(in);                                            \
-  }                                                                 \
-  double gt_##funcName(const std::vector<double> &in) {             \
-    return funcName(in);                                            \
-  }                                                                 \
-  long double gt_##funcName(const std::vector<long double> &in) {   \
-    return funcName(in);                                            \
-  }                                                                 \
 
 GT_DEFINE(distribution)
