+++ sandbox/testinputs/makefile
@@ -1,6 +1,8 @@
 TARGET        = testinputs
 
-CC            = g++
+OPT_FLAGS    ?= -O3 -funsafe-math-optimizations
+
+CXX           = g++
 COMMON_FLAGS += -std=c++14
 COMMON_FLAGS += -g
 COMMON_FLAGS += -Werror
@@ -8,15 +10,14 @@ COMMON_FLAGS += -Wall
 COMMON_FLAGS += -Wextra
 COMMON_FLAGS += -Wstrict-aliasing=0
 UNOPT_FLAGS   = -O0
-OPT_FLAGS    += -O2
-OPT_FLAGS    += -funsafe-math-optimizations
 LINK_FLAGS    =
 DEP_FLAGS     = -MD -MF $*.d
 
 SOURCE        = $(wildcard *.cpp)
-UNOPT_OBJ    += groundtruth.unopt.o
-OPT_OBJ      += test.opt.o
-OPT_OBJ      += main.opt.o
+UNOPT_SOURCE  = groundtruth.cpp
+OPT_SOURCE    = $(filter-out $(UNOPT_SOURCE),$(SOURCE))
+UNOPT_OBJ     = $(UNOPT_SOURCE:%.cpp=%.unopt.o)
+OPT_OBJ       = $(OPT_SOURCE:%.cpp=%.opt.o)
 OBJ           = $(UNOPT_OBJ) $(OPT_OBJ)
 DEPS          = $(SOURCE:%.cpp=%.d)
 
@@ -37,13 +38,13 @@ distclean: clean
 	rm -f $(TARGET)
 
 $(TARGET): $(OBJ) makefile
-	$(CC) -o $(TARGET) $(LINK_FLAGS) $(OBJ)
+	$(CXX) -o $(TARGET) $(LINK_FLAGS) $(OBJ)
 
 %.unopt.o: %.cpp makefile
-	$(CC) -c $(COMMON_FLAGS) $(UNOPT_FLAGS) $(DEP_FLAGS) -o $@ $<
+	$(CXX) -c $(COMMON_FLAGS) $(UNOPT_FLAGS) $(DEP_FLAGS) -o $@ $<
 
 %.opt.o: %.cpp makefile
-	$(CC) -c $(COMMON_FLAGS) $(OPT_FLAGS) $(DEP_FLAGS) -o $@ $<
+	$(CXX) -c $(COMMON_FLAGS) $(OPT_FLAGS) $(DEP_FLAGS) -o $@ $<
 
 .PRECIOUS: %.d
 -include $(DEPS)
