+++ sandbox/testinputs/test.h
@@ -2,12 +2,8 @@
 #define TEST_H
 
 #include "functions.hpp"
+#include "macros.h"
 #include <vector>
-
-#define TST_DECLARE(funcName)                                   \
-  float tst_##funcName(const std::vector<float>&);              \
-  double tst_##funcName(const std::vector<double>&);            \
-  long double tst_##funcName(const std::vector<long double>&);  \
 
 TST_DECLARE(distribution)
